server/requirements.txt
aniso8601==2.0.0
arrow==0.12.1
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
bottlenose==1.1.8
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
freeze==1.0.10
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
marshmallow==2.15.0
pkg-resources==0.0.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
sh==1.12.14
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
Werkzeug==0.14.1
